requirements.txt

django
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary

